requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chromadb==0.3.26
click==8.1.7
clickhouse-connect==0.8.8
coloredlogs==15.0.1
dataclasses-json==0.5.14
duckdb==1.1.3
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hnswlib==0.8.0
httptools==0.6.4
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.0.200
langchainplus-sdk==0.0.20
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.14.3
numexpr==2.10.2
numpy==1.26.4
onnxruntime==1.20.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.7.0
packaging==23.2
pandas==2.2.3
Pillow==9.5.0
posthog==3.7.4
propcache==0.2.0
protobuf==4.25.5
pulsar-client==3.5.0
pyarrow==18.1.0
pydantic==1.10.19
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
streamlit==1.24.0
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.0
websockets==14.1
yarl==1.18.0
zipp==3.21.0
zstandard==0.23.0
